refacor model retrievers, added vehicle retrieval function
