+++ PMD_Backend/interfaces/Retrievable.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PMD_Backend.interfaces
+{
+    public interface Retrievable
+    {
+        public string Retrieve();
+    }
+}
